增加普通的调用算法记录，保留记录状态

+++ label_studio/projects/prompt_api.py
@@ -24,6 +24,8 @@ from db_ml.services import generate_uuid
 from db_ml.services import redis_set_json, redis_get_json
 from db_ml.services import get_project_labels
 from db_ml.services import cut_task_to_model
+from db_ml.services import query_last_record
+from db_ml.services import create_model_record
 from db_ml.listener_result import thread_read_redis_celery_result
 
 
@@ -54,49 +56,45 @@ class PromptLearning(APIView):
     #     return Response(result, status=resp_status)
 
     def post(self, request, *args, **kwargs):
+        """
+        提示学习的触发入口
+        """
         params = request.data
         print('params', params)
         project_id = params['project_id']
         model_id = params.get('model_id', None)
+
+        if query_last_record(project_id):
+            return Response(status=400, data=dict(msg='Project is running model.'))
+
         try:
             # 获取templates
-            template_list = PromptTemplates.objects.filter(
-                project_id=project_id
-            ).values()
+            template_list = PromptTemplates.objects.filter(project_id=project_id).values()
             # print('template_list', template_list)
             templates = [item['template'] for item in template_list]
             # 获取tasks
             tasks = Task.objects.filter(project_id=project_id).values()
-            redis_key = generate_redis_key('prompt', str(project_id))
             if not tasks:
-                return Response(
-                    status=status.HTTP_400_BAD_REQUEST,
-                    data=dict(msg='Invalid project id')
-                )
+                return Response(status=400, data=dict(msg='Invalid project id'))
 
             if PromptResultDraft.objects.filter(project_id=project_id).exists():
                 PromptResultDraft.objects.filter(project_id=project_id).delete()
 
-            p_state = redis_get_json(redis_key)
-            if p_state and p_state.get('state') == AlgorithmState.ONGOING:
-                return Response(
-                    status=status.HTTP_400_BAD_REQUEST,
-                    data=dict(msg='Project is running prompt.')
-                )
+            record_status, record = create_model_record(model_id, project_id, request.user)
+            if not record_status:
+                return Response(status=400, data=dict(msg='Invalid model id.'))
 
-            # print('tasks', tasks)
             # 清空project_id对应的PromptResult表
             c = PromptResult.objects.filter(project_id=project_id).all()
             if len(c):
                 save_raw_data(c, PromptResultDraft, PROMPT_BACKUP_FIELDS)
                 PromptResult.objects.filter(project_id=project_id).delete()
 
-            task_data = []
             project = Project.objects.filter(id=project_id).first()
-            _uuid = generate_uuid('prompt', project_id)
-            state, result, total_count = None, None, 0
-            total_count = len(tasks)
+            _uuid = generate_uuid('prompt', record.id)
+            state, result = None, None
             for sub_query in cut_task_to_model(tasks):
+                task_data = []
                 for task in sub_query:
                     dialogue = task.get('data', {}).get('dialogue', [])
                     task_data.append(dict(
@@ -124,22 +122,11 @@ class PromptLearning(APIView):
                         result = '项目未设置标签'
 
                 if not state:
-                    return Response(
-                        status=status.HTTP_400_BAD_REQUEST,
-                        data={'status': 1, 'error': result}
-                    )
+                    return Response(status=status.HTTP_400_BAD_REQUEST, data={'status': 1, 'error': result})
 
             result = {'status': 0, 'error': ''}
             resp_status = status.HTTP_200_OK
-            redis_state = dict(
-                state=AlgorithmState.ONGOING,
-                total=total_count,
-                project_id=project_id,
-                model_id=model_id,
-                username=request.user.username,
-            )
-            redis_set_json(redis_key, redis_state)
-            thread_read_redis_celery_result(project_id, 'prompt')
+            thread_read_redis_celery_result(project_id, 'prompt', record)
         except Exception as e:
             result = {'status': 1, 'error': str(e)}
             resp_status = status.HTTP_500_INTERNAL_SERVER_ERROR
